requirements.txt
colorlog==6.7.0
Django==4.2.7
django-cors-headers==4.3.0
djangorestframework==3.14.0
drf-yasg==1.21.5
gunicorn==20.1.0
PyJWT==2.4.0
django-filter==23.2
django-redis==5.4.0
git+https://github.com/wokengineers/wokengineers-django.git@2ce0a7b

